requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
importlib-resources==5.10.0
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
zipp==3.9.0
